+++ music_album.rb
@@ -1,28 +1,23 @@
-require_relative 'item'
+require_relative 'item.rb'
 
-class Book < Item
-  attr_accessor :publisher, :cover_state
-
-  def initialize(label, genre, author, publisher, cover_state, published_date)
-    super(label, genre, author, published_date)
-    @publisher = publisher
-    @cover_state = cover_state
-    move_to_archive
+class MusicAlbum < Item
+  def initialize(label = nil, genre = nil, author = nil, publish_date = nil, on_spotify: false)
+    super(label, genre, author, publish_date)
+    @on_spotify = on_spotify
   end
 
-  def can_be_archived?
-    archived
+  def can_be_archived?()
+    @on_spotify && archived
   end
 
   def to_json(_options = {})
     {
-      type: 'Book',
+      type: 'MusicAlbum',
       label: @label,
       genre: @genre,
       author: @author,
-      publisher: @publisher,
-      cover_state: @cover_state,
-      publish_date: @published_date
+      publish_date: @publish_date,
+      on_spotify: @on_spotify
     }.to_json
   end
 end
